+++ src/main/java/com/baichen/rpc/breaker/CircuitBreaker.java
@@ -1,0 +1,14 @@
+package com.baichen.rpc.breaker;
+
+import com.baichen.metrics.MetricsData;
+
+public interface CircuitBreaker {
+
+    boolean allowRequest();
+
+    void recordRpc(MetricsData metricsData);
+
+    enum State {
+        CLOSED, OPEN, HALF_OPEN
+    }
+}