+++ src/main/java/com/baichen/rpc/provider/AddImpl.java
@@ -9,7 +9,7 @@ import java.util.concurrent.locks.LockSupport;
 public class AddImpl implements Add {
     @Override
     public int add(int a, int b) {
-//        LockSupport.parkNanos(TimeUnit.SECONDS.toNanos(4));
+        LockSupport.parkNanos(TimeUnit.SECONDS.toNanos(4));
         return a + b;
     }