+++ src/main/java/com/baichen/rpc/provider/ProviderApp.java
@@ -16,6 +16,8 @@ public class ProviderApp {
         providerProperties.setServiceRegistryConfig(serviceRegistryConfig);
         providerProperties.setHost("127.0.0.1");
         providerProperties.setPort(9091);
+        providerProperties.setGlobalLimit(1000);
+        providerProperties.setServiceLimit(300);
         ProviderServer providerServer = new ProviderServer(providerProperties);
         providerServer.register(Add.class, new AddImpl());
         providerServer.start();